[hotfix] Rename rolling.policies to rollingpolicies in file sink packages
